+++ paws.common/NEWS.md
@@ -3,6 +3,7 @@
 * add pagination StopOnSameToken option (#721) aligns with aws-sdk-js-v3 implementation (https://github.com/aws/aws-sdk-js-v3/releases/tag/v3.78.0). Thanks to @wlandau for raising error in `paginate`.
 * tidy up internal function `jmespath_index`
 * fix `aws-global` region when resolving endpoint (#730). Thanks to @atheriel for identifying the issue.
+* fix default region for service s3 (#730). Thanks to @atheriel for identifying the issue.
 
 # paws.common 0.6.4
 * ensure xml build structure is correctly flattened (#597)